dcbba2orb works on Intel for live data!
After massaging the Orb header addition routines, dcbba2orb is working
with live data from a test data concentrator.

A number of elog_complain calls were converted to elog_log so that only
selected portions of the program will actually output an error message,
in keeping with some of the elog guidelines.