Add Intel processor compatibility
Ensure data is output in Big-Endian order regardless of host byte order.
Original code was written for use on a SPARC or PowerPC which is native
Big-Endian.

Also, ensure that we zero out the unused1 field in the Reel header to
cut down on spurious values that were present on all architectures.

Finally, quick cosmetic white space at end of line cleanup