Set fixed trace length and UTC timebasis code
For SEG-Y Ver 1, there is now a field that allows you to specify that
the traces are all fixed length. This allows faster random access
throughout the file, and is a good idea to be set. This field is in an
unused block of the 1975 SEY-Y standard so it should be safe to set it
for all versions of SEG-Y

Antelope actually records data in UTC not GMT, so for SEG-Y REV1 we will
say that our timebasis is UTC (4) instead of GMT (2). Older versions of
SEG-Y will continue to use timebasis 2.