Output SEG-Y version 1 with text in header block
Added -V option to specify the output format version.

Actually output a Textual Header Block if using SEG-Y classic or Rev 1.
Like before, nothing is output for Seismic Unix support. Added some routines
to spit out EBCDIC for SEG-Y "classic". SEG-Y Rev 1 will always use ASCII for
the header block.

A description blurb can be placed in the first record of the text header.
It is limited to 76 characters due to the file format. It can be controlled
either by the -d option or the text_header_description parameter in the
parameter file

If data is output in rev1, don't set non-standard trace header fields
(the PASSCAL extensions) that conflict with the Rev1 standard. This
includes ignoring the request to use a 32-bit sample counter.

For SEG-Y Ver 1, there is now a field that allows you to specify that
the traces are all fixed length. This allows faster random access
throughout the file, and is a good idea to be set. This field is in an
unused block of the 1975 SEY-Y standard so it should be safe to set it
for all versions of SEG-Y

Antelope actually records data in UTC not GMT, so for SEG-Y REV1 we will
say that our timebasis is UTC (4) instead of GMT (2). Older versions of
SEG-Y will continue to use timebasis 2.

Force fixed length trace flag to true. This was part of a big unused block in
SEGY version 0. Since db2segy always outputs fixed length traces with zero
padding, it should not be an issue to force this bit to true.

added new parameter prefer_decimal_degrees which allows you to use
decimal degrees in Rev1 output. It's not on by default,
but it's probably best practice.

Set timebasis to UTC for SEGY Rev1. The SEG-Y Rev 0 standard didn't define UTC
as a code, and used GMT.  Antelope always records in UTC (which has the leap
second corrections applied), so we should set the time code to UTC if we have
it available to us.

Converted many diagnostic calls to fprintf over to elog.

Extended the header file with comments and byte offsets for most fields

LLVM found some code issues with default variable assignments in the
header initialization routines

Converted many fields to use defined symbolic values instead of numbers

Provide a sane default scale value when using decimal degrees and
looking in the database tables. Output a diagnostic message if we're
using a user-supplied scale factor that causes a loss of precision

Add diagnostic output for coordinate scale values depending on the
output mode

Updated example parameter file with new options, more descriptions on
usage, and added pf_revision_time

Extensively revised man page

Fixed up numerous in-line code comments